requirements.txt
ssdeep==3.4
PyExifTool==0.1.1
